requirements.txt
streamlit
PyMuPDF
pdf2image
opencv-python-headless
pytesseract
Pillow
numpy
rapidfuzz>=3.0
fpdf2>=2.7.0
reportlab